Move useService to TokenContext.ts
